refactor: split runner and test modules Signed-off-by: Gregor Zeitlinger <gregor.zeitlinger@grafana.com>
